Bus Rule 할당 테스트 코드 작성

+++ pcb/src/main/java/com/exam/ddd/domain/LayoutRule.java
@@ -2,5 +2,5 @@ package com.exam.ddd.domain;
 
 public interface LayoutRule {
     String getType();
-    Long getParameter();
+    Integer getParameter();
 }
